Move ajax files into signup app and remove unused windowLocationUtil
