--- frontend/src/lib/windowLocationUtil.js
@@ -1,3 +1,0 @@
-export default {
-  getQueryParameters: () => window.location.search,
-};
